+++ internal/provider/cluster_data_source.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 
+	"github.com/hashicorp/terraform-plugin-framework/attr"
 	"github.com/hashicorp/terraform-plugin-framework/datasource"
 	"github.com/hashicorp/terraform-plugin-framework/datasource/schema"
 	"github.com/hashicorp/terraform-plugin-framework/types"
@@ -33,10 +34,18 @@ type ClusterDataSource struct {
 type ClusterDataSourceModel struct {
 	// ConfigurableAttribute types.String `tfsdk:"configurable_attribute"`
 	// ID                    types.String `tfsdk:"id"`
-	CxProfileName types.String          `tfsdk:"cx_profile_name"`
-	Name          types.String          `tfsdk:"name"`
-	Version       *versionModel         `tfsdk:"version"`
-	Nodes         []NodeDataSourceModel `tfsdk:"nodes"`
+	CxProfileName        types.String          `tfsdk:"cx_profile_name"`
+	Name                 types.String          `tfsdk:"name"`
+	Version              *versionModel         `tfsdk:"version"`
+	Nodes                []NodeDataSourceModel `tfsdk:"nodes"`
+	Contact              types.String          `tfsdk:"contact"`
+	Location             types.String          `tfsdk:"location"`
+	DnsDomains           types.Set             `tfsdk:"dns_domains"`
+	NameServers          types.Set             `tfsdk:"name_servers"`
+	TimeZone             types.Object          `tfsdk:"timezone"`
+	Certificate          types.Object          `tfsdk:"certificate"`
+	NtpServers           types.Set             `tfsdk:"ntp_servers"`
+	ManagementInterfaces types.Set             `tfsdk:"management_interfaces"`
 }
 
 // NodeDataSourceModel describes the data source data model.
@@ -95,6 +104,79 @@ func (d *ClusterDataSource) Schema(ctx context.Context, req datasource.SchemaReq
 					},
 				},
 			},
+			"contact": schema.StringAttribute{
+				Computed:            true,
+				MarkdownDescription: "Contact information. Example: support@company.com",
+			},
+			"location": schema.StringAttribute{
+				Computed:            true,
+				MarkdownDescription: "Location information",
+			},
+			// "password": schema.StringAttribute{
+			// 	Computed:            true,
+			// 	Sensitive:           true,
+			// 	MarkdownDescription: "Password",
+			// },
+			"dns_domains": schema.SetAttribute{
+				ElementType:         types.StringType,
+				Computed:            true,
+				MarkdownDescription: "A list of DNS domains.",
+			},
+			"name_servers": schema.SetAttribute{
+				ElementType:         types.StringType,
+				Computed:            true,
+				MarkdownDescription: "The list of IP addresses of the DNS servers. Addresses can be either IPv4 or IPv6 addresses.",
+			},
+			"timezone": schema.SingleNestedAttribute{
+				Attributes: map[string]schema.Attribute{
+					"name": schema.StringAttribute{
+						Computed:            true,
+						MarkdownDescription: "Time zone",
+					},
+				},
+				Computed:            true,
+				MarkdownDescription: "Time zone",
+			},
+			"certificate": schema.SingleNestedAttribute{
+				Attributes: map[string]schema.Attribute{
+					"id": schema.StringAttribute{
+						Computed: true,
+					},
+				},
+				Computed:            true,
+				MarkdownDescription: "Certificate",
+			},
+			"ntp_servers": schema.SetAttribute{
+				ElementType:         types.StringType,
+				Computed:            true,
+				MarkdownDescription: "Host name, IPv4 address, or IPv6 address for the external NTP time servers.",
+			},
+			"management_interfaces": schema.SetNestedAttribute{
+				NestedObject: schema.NestedAttributeObject{
+					Attributes: map[string]schema.Attribute{
+						"ip": schema.SingleNestedAttribute{
+							Attributes: map[string]schema.Attribute{
+								"address": schema.StringAttribute{
+									Computed:            true,
+									MarkdownDescription: "IP address",
+								},
+							},
+							Computed:            true,
+							MarkdownDescription: "IP address",
+						},
+						"name": schema.StringAttribute{
+							Computed:            true,
+							MarkdownDescription: "Name",
+						},
+						"id": schema.StringAttribute{
+							Computed:            true,
+							MarkdownDescription: "ID",
+						},
+					},
+				},
+				Computed:            true,
+				MarkdownDescription: "A list of network interface",
+			},
 		},
 	}
 }
@@ -148,6 +230,111 @@ func (d *ClusterDataSource) Read(ctx context.Context, req datasource.ReadRequest
 	data.Version = &versionModel{
 		Full: types.StringValue(cluster.Version.Full),
 	}
+	data.Contact = types.StringValue(cluster.Contact)
+	data.Location = types.StringValue(cluster.Location)
+
+	// dns domains
+	elements := []attr.Value{}
+	for _, dnsDomain := range cluster.DnsDomains {
+		elements = append(elements, types.StringValue(dnsDomain))
+	}
+	setValue, diags := types.SetValue(types.StringType, elements)
+	resp.Diagnostics.Append(diags...)
+	data.DnsDomains = setValue
+
+	//name servers
+	elements = []attr.Value{}
+	for _, nameServer := range cluster.NameServers {
+		elements = append(elements, types.StringValue(nameServer))
+	}
+	setValue, diags = types.SetValue(types.StringType, elements)
+	resp.Diagnostics.Append(diags...)
+	data.NameServers = setValue
+	// time zone
+	elementTypes := map[string]attr.Type{
+		"name": types.StringType,
+	}
+	objectElements := map[string]attr.Value{
+		"name": types.StringValue(cluster.TimeZone.Name),
+	}
+	objectValue, diags := types.ObjectValue(elementTypes, objectElements)
+	if diags.HasError() {
+		resp.Diagnostics.Append(diags...)
+		return
+	}
+	data.TimeZone = objectValue
+
+	// certificate
+	elementTypes = map[string]attr.Type{
+		"id": types.StringType,
+	}
+	objectElements = map[string]attr.Value{
+		"id": types.StringValue(cluster.ClusterCertificate.ID),
+	}
+	objectValue, diags = types.ObjectValue(elementTypes, objectElements)
+	if diags.HasError() {
+		resp.Diagnostics.Append(diags...)
+		return
+	}
+	data.Certificate = objectValue
+
+	// ntp servers
+	elements = []attr.Value{}
+	for _, ntpServer := range cluster.NtpServers {
+		elements = append(elements, types.StringValue(ntpServer))
+	}
+	setValue, diags = types.SetValue(types.StringType, elements)
+	resp.Diagnostics.Append(diags...)
+	data.NtpServers = setValue
+
+	// management interfaces
+	setElements := []attr.Value{}
+	for _, mgmInterface := range cluster.ManagementInterfaces {
+		nestedElementTypes := map[string]attr.Type{
+			"address": types.StringType,
+		}
+		nestedVolumeElements := map[string]attr.Value{
+			"address": types.StringValue(mgmInterface.IP.Address),
+		}
+		originVolumeObjectValue, diags := types.ObjectValue(nestedElementTypes, nestedVolumeElements)
+		if diags.HasError() {
+			resp.Diagnostics.Append(diags...)
+			return
+		}
+
+		elementTypes := map[string]attr.Type{
+			"ip":   types.ObjectType{AttrTypes: nestedElementTypes},
+			"name": types.StringType,
+			"id":   types.StringType,
+		}
+		elements := map[string]attr.Value{
+			"ip":   originVolumeObjectValue,
+			"name": types.StringValue(mgmInterface.Name),
+			"id":   types.StringValue(mgmInterface.ID),
+		}
+		objectValue, diags := types.ObjectValue(elementTypes, elements)
+		if diags.HasError() {
+			resp.Diagnostics.Append(diags...)
+			return
+		}
+		setElements = append(setElements, objectValue)
+	}
+
+	setValue, diags = types.SetValue(types.ObjectType{
+		AttrTypes: map[string]attr.Type{
+			"ip": types.ObjectType{AttrTypes: map[string]attr.Type{
+				"address": types.StringType,
+			}},
+			"name": types.StringType,
+			"id":   types.StringType,
+		},
+	}, setElements)
+	if diags.HasError() {
+		resp.Diagnostics.Append(diags...)
+		return
+	}
+
+	data.ManagementInterfaces = setValue
 
 	nodes, err := interfaces.GetClusterNodes(errorHandler, *client)
 	if err != nil {